Flowchart refactor
